Refactor some plotting functions.
`fgraph`: remove the `range` and `size` parameters. Let the ggplot
handles the target range and size of the plot.

`graph2d`: remove the `size` parameters. Let the ggplot
handles the target size of the plot.remove the

Reduce the parameters usage for `fgraph-int`, `bgraph-int`, and
`graph2d`. Let the ggplot handles the target range and size of the plot.